+++ pykeg/contrib/twitter/views.py
@@ -19,11 +19,9 @@
 from django.contrib import messages
 from django.core.urlresolvers import reverse
 from django.shortcuts import redirect
-from socialregistration.clients.oauth import OAuthError
 from pykeg.web.decorators import staff_member_required
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render_to_response
-from django.template import RequestContext
+from django.shortcuts import render
 
 from httplib2 import HttpLib2Error
 
@@ -37,7 +35,7 @@ SESSION_KEY_USER_TWITTER = 'user-twitter'
 
 @staff_member_required
 def admin_settings(request, plugin):
-    context = RequestContext(request)
+    context = {}
 
     consumer_key, consumer_secret = plugin.get_credentials()
     initial = {
@@ -74,7 +72,7 @@ def admin_settings(request, plugin):
                     oauth_token = profile.get('oauth_token')
                     oauth_token_secret = profile.get('oauth_token_secret')
                     tasks.send_tweet(consumer_key, consumer_secret, oauth_token, oauth_token_secret,
-                        tweet, image_url=None)
+                                     tweet, image_url=None)
                     messages.success(request, 'Tweet sent')
                 else:
                     messages.error(request, 'There was a problem sending tweet')
@@ -86,12 +84,12 @@ def admin_settings(request, plugin):
     context['settings_form'] = settings_form
     context['tweet_form'] = tweet_form
 
-    return render_to_response('contrib/twitter/admin_settings.html', context_instance=context)
+    return render(request, 'contrib/twitter/admin_settings.html', context=context)
 
 
 @login_required
 def user_settings(request, plugin):
-    context = RequestContext(request)
+    context = {}
     user = request.user
 
     consumer_key, consumer_secret = plugin.get_credentials()
@@ -110,7 +108,7 @@ def user_settings(request, plugin):
     context['profile'] = plugin.get_user_profile(user)
     context['settings_form'] = settings_form
 
-    return render_to_response('contrib/twitter/twitter_user_settings.html', context_instance=context)
+    return render(request, 'contrib/twitter/twitter_user_settings.html', context=context)
 
 
 @staff_member_required
@@ -121,36 +119,23 @@ def site_twitter_redirect(request):
         messages.success(request, 'Removed Twitter account.')
         return redirect('kegadmin-plugin-settings', plugin_name='twitter')
 
-    plugin = request.plugins['twitter']
-
-    client = plugin.get_client()
-    url = request.build_absolute_uri(reverse('plugin-twitter-site_twitter_callback'))
-    client.set_callback_url(url)
-
-    return do_redirect(request, client, 'kegadmin-plugin-settings', SESSION_KEY_SITE_TWITTER)
+    return do_redirect(request, 'plugin-twitter-site_twitter_callback', 'kegadmin-plugin-settings', SESSION_KEY_SITE_TWITTER)
 
 
 @staff_member_required
 def site_twitter_callback(request):
     plugin = request.plugins.get('twitter')
-    client = request.session.get(SESSION_KEY_SITE_TWITTER)
-
-    if not client:
-        messages.error(request, 'Lost OAuth session')
+    try:
+        resource_owner_key, resource_owner_secret = request.session.get(SESSION_KEY_SITE_TWITTER, [])
+    except ValueError:
+        messages.error(request, 'Lost session, please try again.')
     else:
+        client = plugin.get_client()
         del request.session[SESSION_KEY_SITE_TWITTER]
-        try:
-            token = client.complete(dict(request.GET.items()))
-        except KeyError:
-            messages.error(request, 'Session expired.')
-        except OAuthError, error:
-            messages.error(request, str(error))
-        else:
-            user_info = client.get_user_info()
-            plugin = request.plugins.get('twitter')
-            plugin.save_site_profile(token.key, token.secret, user_info['screen_name'],
-                int(user_info['user_id']))
-            messages.success(request, 'Successfully linked to @%s' % user_info['screen_name'])
+        token, token_secret = client.handle_authorization_callback(resource_owner_key, resource_owner_secret, request)
+        user_info = client.get_user_info(token, token_secret)
+        plugin.save_site_profile(token, token_secret, user_info.screen_name, user_info.id)
+        messages.success(request, 'Successfully linked to @%s' % user_info.screen_name)
 
     return redirect('kegadmin-plugin-settings', plugin_name='twitter')
 
@@ -163,58 +148,48 @@ def user_twitter_redirect(request):
         messages.success(request, 'Removed Twitter account.')
         return redirect('account-plugin-settings', plugin_name='twitter')
 
-    plugin = request.plugins['twitter']
-    client = plugin.get_client()
-    url = request.build_absolute_uri(reverse('plugin-twitter-user_twitter_callback'))
-    client.set_callback_url(url)
-
-    return do_redirect(request, client, 'account-plugin-settings', SESSION_KEY_USER_TWITTER)
+    return do_redirect(request, 'plugin-twitter-user_twitter_callback', 'account-plugin-settings', SESSION_KEY_USER_TWITTER)
 
 
 @login_required
 def user_twitter_callback(request):
     plugin = request.plugins['twitter']
-    client = request.session.get(SESSION_KEY_USER_TWITTER)
-
-    if not client:
-        messages.error(request, 'Lost OAuth session')
+    try:
+        resource_owner_key, resource_owner_secret = request.session.get(SESSION_KEY_USER_TWITTER, [])
+    except ValueError:
+        messages.error(request, 'Lost session, please try again.')
     else:
+        client = plugin.get_client()
         del request.session[SESSION_KEY_USER_TWITTER]
-        try:
-            token = client.complete(dict(request.GET.items()))
-        except KeyError:
-            messages.error(request, 'Session expired.')
-        except OAuthError, error:
-            messages.error(request, str(error))
-        else:
-            user_info = client.get_user_info()
-            plugin = request.plugins.get('twitter')
-            plugin.save_user_profile(request.user, token.key, token.secret,
-                user_info['screen_name'], int(user_info['user_id']))
-            messages.success(request, 'Successfully linked to @%s' % user_info['screen_name'])
+        token, token_secret = client.handle_authorization_callback(resource_owner_key, resource_owner_secret, request)
+        user_info = client.get_user_info(token, token_secret)
+        plugin.save_user_profile(request.user, token, token_secret, user_info.screen_name, user_info.id)
+        messages.success(request, 'Successfully linked to @%s' % user_info.screen_name)
 
     return redirect('account-plugin-settings', plugin_name='twitter')
 
 
-def do_redirect(request, client, next_url_name, session_key):
+def do_redirect(request, callback_url_name, next_url_name, session_key):
     """Common redirect method, handling any incidental errors.
 
     Args:
         request: the incoming request
-        client: the TwitterClient context_instance
+        client: the TwitterClient context
         next_url_name: Django URL name to redirect to upon error
 
     Returns:
         A redirect response, either to the OAuth redirect URL
         or to `next_url_name` on error.
     """
-    request.session[session_key] = client
+    callback_url = request.build_absolute_uri(reverse(callback_url_name))
+    client = request.plugins['twitter'].get_client()
+
+    url, resource_owner_key, resource_owner_secret = client.get_redirect_url(callback_url)
+    request.session[session_key] = (resource_owner_key, resource_owner_secret)
+
     try:
-        return redirect(client.get_redirect_url())
-    except OAuthError, e:
-        messages.error(request, 'Error: %s' % str(e))
-        return redirect(next_url_name, plugin_name='twitter')
-    except HttpLib2Error, e:
+        return redirect(url)
+    except HttpLib2Error as e:
         # This path can occur when api.twitter.com is unresolvable
         # or unreachable.
         messages.error(request, 'Twitter API server not available. Try again later. (%s)' % str(e))