+++ pykeg/contrib/untappd/untappd_test.py
@@ -39,12 +39,16 @@ class UntappdTests(TransactionTestCase):
         self.fake_plugin_registry = {'foursquare': self.fsq}
 
         self.plugin = plugin.UntappdPlugin(datastore=self.datastore,
-            plugin_registry=self.fake_plugin_registry)
+                                           plugin_registry=self.fake_plugin_registry)
         self.user = models.User.objects.create(username='untappd_test')
         self.backend = get_kegbot_backend()
         self.tap = self.backend.create_tap('Test Tap', 'test.flow0')
-        self.keg = self.backend.start_keg(tap=self.tap, beverage_name='Test Beer',
-            beverage_type='beer', producer_name='Test Producer', style_name='Test Style')
+        self.keg = self.backend.start_keg(
+            tap=self.tap,
+            beverage_name='Test Beer',
+            beverage_type='beer',
+            producer_name='Test Producer',
+            style_name='Test Style')
         self.beverage = self.keg.type
         self.beverage.untappd_beer_id = '9876'
         self.beverage.save()
@@ -64,16 +68,20 @@ class UntappdTests(TransactionTestCase):
         self.assertEqual('fake-token', self.plugin.get_user_token(self.user))
 
         fake_drink = models.Drink.objects.create(keg=self.keg, volume_ml=1000, ticks=1000,
-            user=self.user, time=timezone.now(), shout='Hello')
-        fake_event = models.SystemEvent.objects.create(kind=models.SystemEvent.DRINK_POURED,
-            drink=fake_drink, user=self.user, keg=self.keg, time=fake_drink.time)
+                                                 user=self.user, time=timezone.now(), shout='Hello')
+        fake_event = models.SystemEvent.objects.create(
+            kind=models.SystemEvent.DRINK_POURED,
+            drink=fake_drink,
+            user=self.user,
+            keg=self.keg,
+            time=fake_drink.time)
 
         with patch('pykeg.contrib.untappd.tasks.untappd_checkin.delay') as mock_checkin:
             self.plugin.handle_new_events([fake_event])
             mock_checkin.assert_called_with('fake-token', '9876', 'UTC', shout='Hello',
-                foursquare_client_id=None,
-                foursquare_client_secret=None,
-                foursquare_venue_id=None)
+                                            foursquare_client_id=None,
+                                            foursquare_client_secret=None,
+                                            foursquare_venue_id=None)
 
     def test_drink_poured_with_foursquare(self):
         fsq_settings = self.fsq.get_site_settings_form()
@@ -87,14 +95,23 @@ class UntappdTests(TransactionTestCase):
         self.plugin.save_user_token(self.user, 'fake-token')
         self.assertEqual('fake-token', self.plugin.get_user_token(self.user))
 
-        fake_drink = models.Drink.objects.create(keg=self.keg, volume_ml=1000, ticks=1000,
-            user=self.user, time=timezone.now(), shout='Hello2')
-        fake_event = models.SystemEvent.objects.create(kind=models.SystemEvent.DRINK_POURED,
-            drink=fake_drink, user=self.user, keg=self.keg, time=fake_drink.time)
+        fake_drink = models.Drink.objects.create(
+            keg=self.keg,
+            volume_ml=1000,
+            ticks=1000,
+            user=self.user,
+            time=timezone.now(),
+            shout='Hello2')
+        fake_event = models.SystemEvent.objects.create(
+            kind=models.SystemEvent.DRINK_POURED,
+            drink=fake_drink,
+            user=self.user,
+            keg=self.keg,
+            time=fake_drink.time)
 
         with patch('pykeg.contrib.untappd.tasks.untappd_checkin.delay') as mock_checkin:
             self.plugin.handle_new_events([fake_event])
             mock_checkin.assert_called_with('fake-token', '9876', 'UTC', shout='Hello2',
-                foursquare_client_id='fake-client-id',
-                foursquare_client_secret='fake-client-secret',
-                foursquare_venue_id='54321')
+                                            foursquare_client_id='fake-client-id',
+                                            foursquare_client_secret='fake-client-secret',
+                                            foursquare_venue_id='54321')