+++ pykeg/core/management/commands/run_workers.py
@@ -17,7 +17,6 @@
 # along with Pykeg.  If not, see <http://www.gnu.org/licenses/>.
 
 import os
-from django.conf import settings
 
 from pykeg.core.management.commands.common import RunnerCommand
 
@@ -39,17 +38,14 @@ class Command(RunnerCommand):
                 os.path.join(logs_dir, 'celery_beat.log'))
         ret = []
 
-        # Pluck out the app_name.  Typically this is 'pykeg', but may be
-        # something else in embedded mode.
-        app_name = settings.SETTINGS_MODULE.split('.')[0]
-        base_cmd = 'celery -A {} worker -l info '.format(app_name)
+        base_cmd = 'celery -A pykeg worker -l info '
 
         ret.append(('celery_default',
-            base_cmd + '-Q default --hostname="default@%h"' + default_log))
+                    base_cmd + '-Q default --hostname="default@%h"' + default_log))
         ret.append(('celery_stats',
-            base_cmd + '-Q stats --concurrency=1 --hostname="stats@%h"' + stats_log))
+                    base_cmd + '-Q stats --concurrency=1 --hostname="stats@%h"' + stats_log))
 
         ret.append(('celery_beat',
-            'celery -A {} beat --pidfile= -l info{}'.format(app_name, beat_log)))
+                    'celery -A pykeg beat --pidfile= -l info{}'.format(beat_log)))
 
         return ret